Create Omnitrix.py
Gives you a set of 10 random aliens, I tend to use my imagination a lot and sometimes find it's nice to be able to get some random unbiased elements to the games, which for instance might be something like this where you start out the adventure with 10 random aliens just like in the shows. This just makes it easier so that you don't have to go to bizarre extreme lengths to get a random selection.